Display absolute path on failed dependency resolve.
When a file path can not be resolved the error message is cleaned
up but the relative path to file is displayed.

In cases where a mistake have been made this path can be very confusing,
like `Cannot resolve dependency './../../../../../../main.js'` for when
a path starts with `/`.

This PR resolves attempted path into an absolute path in hope of
reducing confusion and minimize cognitive overhead of `..` chasing.