Invalidate cache when dependencies included in parent change
Some dependencies are pre-compiled into their parent assets, e.g. stylus imports. But we still need to invalidate the cache if any of those files change.

Now the FSCache looks for deps with `includedInParent: true`, and stores an mtime in the cache file. When reading from the cache, those deps are checked to see if they changed and if so, the parent asset is invalidated.