Changed automatic install to use installPackage helper
I had to add a parameter to installPackage to make saveDev optional, I also cleaned up the install code a bit in the proccess.

I also fixed a bug where installPackage would naively use yarn as long as it found a "yarn.lock" which would cause issues if a user doesn't have yarn but is working on a project that commits its yarn.lock to the repo. So, I used the command-exists package to check first to prevent this issue from happening.