Added back yarn.lock check, improved yarn.lock & project root resolution
I added back the check for yarn.lock but also kept the yarn command check. So now it checks for both the yarn command and a yarn.lock file.

Also found bug where if you are using a flat project structure (ie entry point is not within a sub directory such as src), the 'location' variable would be null and result in yarn.lock never being found.

So, I improved the logic to use the provided dir as a fallback to the config file directory, so if it comes back null dir get's used as the assumed project root and as a result stuff get's installed there and that's where we look for the yarn.lock