Bug fix
I originally had it the other way but right before my last commit I decdided to clean up the code and refactor it this way. Problem is that while it looks cleaner, it throws an error when configFileLocation is null and get's passed into path.dirName as null.

So ya... just moved it back to the original way :)