Fixed Node 6 Compatiblity Issues
I relied on the fact that the latest version of npm also has the `npm add` command (just like yarn), so I was able to make it really cleanly use `add` as the command for both.

Problem is that we also need to support legacy versions of node, and with it, legacy versions of npm (which don't support `npm add`).

So I put it back to how it was originally pretty much... now it checks if it's npm or yarn, and based on that decides if it needs to use 'npm install' or 'yarn add'.