Removed recursive config file search
Before we were using config.resolve() to attempt to find where the nearest config file is, which is then used as the projectRootLocation. The problem is that config.resolve() also searches previous directories which causes oubvious problems since if you have a yarn.lock somewhere in a parent directory, it doesn't nessarely mean its from this project but config.resolve() would still find it. This was a big issue because it would break tests, since it would use Parcel's package.json / yarn.lock and end up installing packages in the Parcel repo root directory rather than inside the test input directory.

Also, there's not really much use for a recursive search here, since all the code that uses install() provides the correct install directory, so we might as well just exect the correct input isntead of expecting the wrong input and then trying to hack together a way to figure out the right one.

I suppose if we do find it nessary to do a reccursive resolve for project config files in the future then we can always just put it back to how it was and then statically compare the path strings to make sure we didn't escape the project directory.