Added back reccursive yarn.lock resolution
Ok lemme explain what happened (in case your wondering why I keep going back and forth with all this config.resolve stuff):

Originally we would just attempt to resolve either package.json OR yarn.lock and then use that as the 'location' (later renamed to projectRootLocation).

This caused a ton of issues, so I got rid of this conig.resolve entierely. 

Then I realized that yarn.lock won't get resolved if it's not directly in the directory, so now here I am adding back config.resolve(), but only for the purpuse of resolving the yarn lock file, and not for any of that project root stuff.