CHANGED CONFIG.RESOLVE() TO CHECK THE DIRECORY BEFORE GOING TO PARENT
Unit tests are passing, so I hope this doesn't break anything, but this fixes a huge subtle bug in the way config.resolve() works.

Originally config.resolve() would set filepath to dirName(filePath) which would implicitly change the filepath to the parent directory, allowing the recursion to work.

However, this is a problem when this statement is on the top, if the initial path passed in is a directory, because then that directory get's skipped.

So I moved it to the bottom, after everything else runs, to make sure that we check the current directory before moving on to the parent (which makes sense)

I'm not sure how this will behave if you pass in a path to a file rather than a directory. By the looks of it, it'll just preform a redundent recursive step (we should probably fix this too)