Fix babel config in package.json
Babel seems to mutate the passed-in config, so if we pass in the
package.babel directly it will mutate the package config, then the next
time it compiles the same file it will have duplicate plugins, etc.

Instead, if loading config from a packageKey, clone the config before
returning so it can't be mutated.