separate updating graph from anywhere and completing a graph's edges
the edges of a graph should not keep processing if an update
to the middle graph might prune those edges. if an update comes in,
stop everything, and start over making sure the graph is updated
before processing the edges. if an update comes in while the graph
is in the middle of updating, just add it to the update queue, and
don't cancel anything