Keep babel env option while transforming
According to https://new.babeljs.io/docs/en/next/babelrc.html#env-environment-option, we can specify a plugin to run in a certain environment. But right now the `env` option will be ignored while parcel is doing the babel transformation.

Parcel will ignore the `env` option in the following `.babelrc` so it's gonna treat that `.babelrc` as an empty one and do nothing.
```js
{
  "env": {
    "development": {
      "plugins": ["babel-plugin-styled-components"]
    }
  }
}
```