No longer use unref in Workers
This was originally added for @parcel/register so processes would not hang,
but it also caused a weird issue where Parcel could potentially exit early
with a clean exit code. It is also not even beneficial for @parcel/register
to use Workers because it is hooking into require which is synchronous and
cannot be parallelized.