Reload css form local hot server
CSS file can be included in HTML file or in HTML template
of a web framework (php or ruby, etc). In this case
the location origing will not match with CSS href and CSS will
not be reloaded. To fix that we check if CSS is served from
the localhost and its port is the same as parcel HMR port