Support <script type="module"> when serving with HMR
Hot Module Replacement was failing when a module type script was present on the
page. Building for development now removes this attribute, allowing HMR to be
used. The attribute is preserved when using the parcel build command.

The nomodule attribute on scripts is also ignored in development to prevent two
scripts from running in browsers without module support.