Reload css form local HMR server (#5061)
* Reload css form local hot server

CSS file can be included in HTML file or in HTML template
of a web framework (php or ruby, etc). In this case
the location origing will not match with CSS href and CSS will
not be reloaded. To fix that we check if CSS is served from
the localhost and its port is the same as parcel HMR port

* Updatde logic to detect if asset served from hmr

* Update hmr-runtime.js

Co-authored-by: Maksim Semenov <maksimsemenov@Maksims-MacBook-Air.local>
Co-authored-by: Jasper De Moor <jasperdemoor@gmail.com>