Implement XOR doubly linked edge list
This is a space optimization; there may be some performance impact
due to the extra time spent scanning the list when removing an edge.

It saves space by eliminating the need to explicitly store the
next and previous edges in each edge object. Instead, it maintains
an XOR of the next and previous edges on each edge, which means
following the links between edges requires knowing the edge that
came before or after.