Reorder resolveOptions() env priority
This change updates the priority of the destructed environment variables in
resolveOptions() so that process.env and initialOptions.env are not overwritten
by the result of loadDotEnv().

This fixes an issue where environment variables passed on the command line were
overwritten by the values in the .env file.

The following now works as expected:

`SOME_VAR=foo npm run build`