Refactor nodes and edges into type maps
This looks like a big change because a lot of stuff moved around.
The edges are currently stored the same way they were before,
but the hash table implementation has been extracted for use
in the new node map.

The node map is where the real change is; instead of just linking
nodes to addresses in the edge map, it now stores multiple sets
of links per node, one for each type of edge.

This allows us to eliminate the need for a type map cache,
as we are now able to look up edges by type in constant time.