Add integration tests and fix resulting bugs.
These tests cover the scenarios in which typescript preprocessing is
required and where no preprocessing is desired. There were a few bugs
discovered as a result. The preprocessor config being set directly on
the transformer config was disabling caching. The preprocessor config
was being set to the function which produces a preprocessor from
svelte-preprocess rather than the result of that function.