fix: force the use of posix path seperators for ts module graphs
When resolving the name of the main module for use in TS module graphs,
ensure we force the usage of posix paths as this is what `typescript`
uses and returns.
This fixes an issue when running on Windows, where an assert would be
thrown as a result of two identical paths using different path
seperators being compared for equality.