Fix dynamic import of modules with inlined imports
When dynamically importing a module that itself has an inlined import
(e.g. with bundle-text) a module loader is added for the inlined bundle,
which references a file that is not generated (since it is inlined).
This invalid reference causes the import to fail at runtime if the fetch
fails before the correct import resolves; it appears to also cause a
TypeError ("Cannot read properties of undefined") in tests.

This fix simply excludes external bundles in getLoaderRuntime that have
a bundleBehavior of inline.