Migrate to swc_core
Using `swc_ecmascript` etc directly is deprecated and the recommendation
is to migrate to swc_core instead as a meta-pacakge.

This change makes that migration - it's mostly a case of updating a lot
of imports, but also as a side-effect of bumping a few versions of SWC
to pick up bugfixes there were some deprecated fields where code changes
were needed.

One major change with the latest SWC is that it currently requires Rust
nightly to build. I've updated the `rust-toolchain` to reflect this, but
I'm not sure if there are any broader implications.