Retry integration tests 2 times on CI
As raised in #9689, integration tests are currently flaky. This makes the use
of a merge queue harder since with merge queues individual builds can't be
retried.

As a temporary work-around this commit adds retries to the integration tests'
mocha invocation.

This is very subpar and we must come-up with a plan to address the tests'
unreliability.