SWC plugin transformer example
This commit adds a parcel transformer that executes SWC.

This is a way that parcel can support arbitrary SWC plugins,
although this might not have good performance, and needs more tweaking
to be usable.

For some reason, the SWC path resolution doesn't work unless the `.swcrc`
references the absolute path of the plugin WASM file.

Where you see `/home/ubuntu/parcel-2/...` on the `.swcrc` that's the
path on my machine.